// src/app/models/egitimkayit.ts
export class egitimKayit {
    uid: string;
    key: string;
    egitimFotoUrl: string;
    egitimFotoAdi: string;
    egitimFile: File;
    egitimAd: string;
    egitimBilgi: string;
    egitimKategori: string;
    egitimFiyat: string;
    egitimKisaBir: string;
    egitimKisaIki: string;
    egitimKisaUc: string;
    egitimKisaDort: string;
    egitimKisaBes: string;
    egitimKisaAlti: string;
    kayTarih: string;
    duzTarih: string;
}