// src/app/models/favoriegitim.ts
export class favoriEgitim {
    key: string;
    uid: string;
    egitimID: string;
    egitimFotoUrl: string;
    egitimAd: string;
    egitimBilgi: string;
    egitimKategori: string;
    egitimFiyat: string;
    egitimKisaBir: string;
    egitimKisaIki: string;
    egitimKisaUc: string;
    egitimKisaDort: string;
    egitimKisaBes: string;
    egitimKisaAlti: string;
    kayTarih: string;
}